html/semantics/forms/the-input-element/hidden.html

<!DOCTYPE html>
<html>
  <head>
    <title>Hidden input element</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
    <link rel="author" title="Kinuko Yasuda" href="mailto:kinuko@chromium.org">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/forms.html#hidden-state-(type=hidden)">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <h1>Hidden input element</h1>
    <div style="display: none">

    <input id="hidden" type="hidden" />
    <input id="hidden_with_value" type="hidden" value="foo" />

    </div>
    <div id="log"></div>
  <script type="text/javascript">

    test(
        function() {
          assert_equals(document.getElementById("hidden").value, "");
          assert_equals(document.getElementById("hidden_with_value").value, "foo");
        }, "Value returns the current value for hidden");

    test(
        function() {
          document.getElementById("hidden").value = "A";
          assert_equals(document.getElementById("hidden").value, "A");
          document.getElementById("hidden").value = "B";
          assert_equals(document.getElementById("hidden").value, "B");
        }, "Setting value changes the current value for hidden");

    test(
        function() {
          assert_equals(document.getElementById("hidden").files, null);
        }, "files attribute must return null for hidden");

    test(
        function() {
          assert_equals(document.getElementById("hidden").valueAsDate, null);
        }, "valueAsDate attribute must return null for hidden");

    test(
        function() {
          assert_equals(document.getElementById("hidden").valueAsNumber, NaN);
        }, "valueAsNumber attribute must return NaN for hidden");

    test(
        function() {
          assert_equals(document.getElementById("hidden").list, null);
        }, "list attribute must return null for hidden");

    test(
        function() {
          var el = document.getElementById("hidden");
          el.value = "1";
          var exception;
          try {
            el.stepDown();
          } catch (e) {
            exception = e;
          }
          assert_equals(el.value, "1");
          assert_equals(exception.name, "InvalidStateError");
        }, "stepDown does not apply for hidden");

    test(
        function() {
          var el = document.getElementById("hidden");
          el.value = "1";
          var exception;
          try {
            el.stepUp();
          } catch (e) {
            exception = e;
          }
          assert_equals(el.value, "1");
          assert_equals(exception.name, "InvalidStateError");
        }, "stepUp does not apply for hidden");

  </script>
  </body>
</html>
